Extract HTTP health check into dedicated crate
