+++ rust/relay/Cargo.toml
@@ -37,8 +37,8 @@ url = "2.4.1"
 serde = { version = "1.0.196", features = ["derive"] }
 trackable = "1.3.0"
 socket2 = "0.5.6"
-axum = { version = "0.7.3", default-features = false, features = ["http1", "tokio"] }
 backoff = "0.4"
+http-health-check = { workspace = true }
 
 [dev-dependencies]
 redis = { version = "0.25.0", default-features = false, features = ["tokio-comp"] }
